### 2018-09-16 #954 @whedon
[PRE REVIEW]: Local clustering
**Submitting author:** @volfpeter (<a href="http://orcid.org/0000-0002-6226-4277">Péter Volf</a>)
**Repository:** <a href="https://github.com/volfpeter/localclustering" target ="_blank">https://github.com/volfpeter/localclustering</a>
**Version:** v0.10
