### 2018-09-16 9f4f75b
trying to make JOSS PDF compilation work, the error message is not too helpful
