### 2018-09-16 #954 @arfon
@volfpeter - please format your `paper.md` YAML header _exactly_ as shown in this [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
