### 2018-09-16 #954 @volfpeter
Hi!
Thanks. I have one question. What affiliation should i set if i'm not part of any organization (i'm self-employed)?
