### 2018-09-16 #954 @pjotrp
Just put in 'Independent' or your company name. I am very supportive of the independents :). We can also start looking for a reviewer. @volfpeter it would be good to put an example/application/figure in your paper to advertise the work. Do you have something that visualizes the algorithm? It is not totally necessary, but I think it will improve the paper. Also you may want to add a paragraph on future plans, if any, and/or how one could expand/modify the library.
