### 2018-09-16 #954 @pjotrp
Oh yes, it would also be good to give it a bit more context with references to prior art/state-of-the-art.
