### 2018-09-16 #954 @volfpeter
@pjotrp Thanks, i'll use independent then :)

I wanted to make the paper as compact as possible. The details (detailed description, evaluation, ideas for future development) are hidden in the documents that i've referenced in the References section and the readme of the project.

The iPython [notebook](https://github.com/volfpeter/localclustering/blob/master/documents/Algorithm%20Analysis%20with%20the%20Spotify%20Related%20Artists%20Graph.ipynb) referenced in the Resources section contains plenty of plots, charts, and tables. I can take one of them and place it somewhere in the paper.

I'll try to compile a list of papers that are related to this topic. Unfortunately i've originally created this a few years ago (recently i only finalised it and published it on github), so my theoretical knowledge is not quite up-to-date.
