### 2018-09-16 32849dd
added affiliation, added References section, referenced the bib file, embedded an example clustering result
