### 2018-09-17 #954 @volfpeter
I've added an example result to the paper.
Now i'm trying to install whedon on my mac, so i can compile the paper locally and not spam the issue with compile errors.
