### 2018-09-17 #954 @arfon
@volfpeter - this PR should fix your paper compilation issues: https://github.com/volfpeter/localclustering/pull/1

Also, you don't appear to be citing any prior work/related materials in your paper yet.
