### 2018-09-17 #954 @volfpeter
@arfon Thanks and yes, the citation list in progress.
