### 2018-09-17 77d29da
added citations to the bib file
