+++ paper.bib
@@ -1,0 +1,61 @@
+@article{,
+    Author = "Blondel, V., Guillaume, J.L., Lambiotte, R., Lefebvre, E.",
+    Title = "Fast unfolding of communities in large networks",
+    Journal = "Journal of Statistical Mechanics: Theory and Experiment",
+    Year = 2008,
+    Volume = "2008",
+    Pages = "P10008"
+}
+
+@article{,
+    Author = "van Dongen, S.",
+    Title = "A cluster algorithm for graphs",
+    Journal = "Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam",
+    Year = 2000,
+    Volume = "Technical Report INS-R0010"
+}
+
+@article{,
+    Author = "Girvan, M., Newman, M.",
+    Title = "Community structure in social and biological networks",
+    Journal = "PNAS",
+    Year = 2002,
+    Volume = "99 (12)",
+    Pages = "7821–7826"
+}
+
+@article{,
+    Author = "Newman, M.",
+    Title = "Finding community structure in networks using the eigenvectors of matrices",
+    Journal = "Physical Review E",
+    Year = 2006,
+    Volume = "74",
+    Pages = "036104"
+}
+
+@article{,
+    Author = "Newman, M., Girvan, M.",
+    Title = "Finding and evaluating community structure in networks",
+    Journal = "Physical Review E",
+    Year = 2004,
+    Volume = "69",
+    Pages = "026113"
+}
+
+@article{,
+    Author = "Pons, P., & Latapy, M.",
+    Title = "Computing communities in large networks using random walks",
+    Journal = "Journal of Graph Algorithms and Applications",
+    Year = 2006,
+    Volume = "10",
+    Pages = "191–218"
+}
+
+@article{,
+    Author = "Schaeffer, S. E.",
+    Title = "Graph clustering",
+    Journal = "Computer Science Review",
+    Year = 2007,
+    Volume = "1",
+    Pages = "27-64"
+}
