### 2018-09-17 4c97fa3
added more citations and references + added an Ideas for future work section
