+++ paper.bib
@@ -1,4 +1,4 @@
-@article{,
+@article{Blondel:2008,
     Author = "Blondel, V., Guillaume, J.L., Lambiotte, R., Lefebvre, E.",
     Title = "Fast unfolding of communities in large networks",
     Journal = "Journal of Statistical Mechanics: Theory and Experiment",
@@ -7,7 +7,7 @@
     Pages = "P10008"
 }
 
-@article{,
+@article{Dongen:2000,
     Author = "van Dongen, S.",
     Title = "A cluster algorithm for graphs",
     Journal = "Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam",
@@ -15,7 +15,7 @@
     Volume = "Technical Report INS-R0010"
 }
 
-@article{,
+@article{Girvan:2002,
     Author = "Girvan, M., Newman, M.",
     Title = "Community structure in social and biological networks",
     Journal = "PNAS",
@@ -24,7 +24,7 @@
     Pages = "7821–7826"
 }
 
-@article{,
+@article{Newman:2006,
     Author = "Newman, M.",
     Title = "Finding community structure in networks using the eigenvectors of matrices",
     Journal = "Physical Review E",
@@ -33,7 +33,7 @@
     Pages = "036104"
 }
 
-@article{,
+@article{Newman:2004,
     Author = "Newman, M., Girvan, M.",
     Title = "Finding and evaluating community structure in networks",
     Journal = "Physical Review E",
@@ -42,8 +42,8 @@
     Pages = "026113"
 }
 
-@article{,
-    Author = "Pons, P., & Latapy, M.",
+@article{Pons:2006,
+    Author = "Pons, P., Latapy, M.",
     Title = "Computing communities in large networks using random walks",
     Journal = "Journal of Graph Algorithms and Applications",
     Year = 2006,
@@ -51,11 +51,29 @@
     Pages = "191–218"
 }
 
-@article{,
+@article{Raghavan:2007,
+    Author = "Raghavan, U. N., Albert, R., Kumara, S.",
+    Title = "Near linear time algorithm to detect community structures in large-scale networks",
+    Journal = "Physical Review E",
+    Year = 2007,
+    Volume = "76",
+    Pages = "036106"
+}
+
+@article{Schaeffer:2007,
     Author = "Schaeffer, S. E.",
     Title = "Graph clustering",
     Journal = "Computer Science Review",
     Year = 2007,
     Volume = "1",
     Pages = "27-64"
+}
+
+@article{Zachary:1977,
+    Author = "Zachary, W.",
+    Title = "An information flow model for conflict and fission in small groups",
+    Journal = "Journal of Anthropological Research",
+    Year = 1977,
+    Volume = "33",
+    Pages = "452--473"
 }
