### 2018-09-17 #954 @volfpeter
Dear @pjotrp and @DannyArends,
I've added the citations to the paper and a short Ideas for future work section. I think (hope) it is ready for review.
