### 2018-09-17 #960 @volfpeter
@pjotrp I've gone through the checklist. I think everything is fine, but here are a couple of notes:
- Version: The latest release is v0.10, the code itself has not changed since that release, but obviously the paper is not included in that release. If the published release must contain the paper, then i will have to update the version.
- Automated tests: There are only manual "tests" in the form of the iPython notebook and `demo.py`. I hope it will be enough.
- References: Do i need to look up the DOI of all citations? If so, then i will get to it.
