### 2018-09-18 #960 @pjotrp
Just create a new release when we are ready. Tests should be fine. DOIs, please add.

@DannyArends you can start review. Thanks!
