### 2018-09-18 #960 @DannyArends
Ticked all the boxes I could, in general the software is in a good state. However how to use the library, what the advantages are, and the use-cases of this clustering algorithm compared to already available methods remains a mystery to me.

I would advice the author to additionally take a close look at the language used in the paper (although this is outside of the scope of the review) I added some comments in the additional section below.

Furthermore I was not able to install the software yet, I will do that after the issues listed below are taken care of.

Functionality - Installation
---
The software only provides a short list of dependencies, no clear install guide is provided.
Please provide an install guide outlining the steps needed to be taken by a new user
to install the software, even if this is a simple as:

```
    pip install graphscraper
    git clone https://github.com/volfpeter/localclustering.git
```
Then what do I need to do ? Do I need to include the library ? does it need compilation/installation ? Do I need to update my path variable ?

Additionally, it is not specified if this requires a certain OS, is this windows, linux, macOSX compatible ? has it been tested on multiple platforms ?

Documentation - A statement of need
---
No statement of need is provided, please make clear in the readme for whom the
software in this repository is useful: e.g. is it aimed at biologist, mathematicians, physics

Documentation - Installation instructions
---
There is no clear list of dependencies, and dependencies for the library, the demo and the python notebook are different. Furthermore, it is unclear which version of "GraphScraper library" is required, since the readme file links to the pypi version (https://pypi.python.org/pypi/graphscraper) as well as the authors fork of the same software: (https://github.com/volfpeter/graphscraper).

A solution would be to split up the dependencies section into three parts outlining the requirements for: Library, Demo and IPython notebook, and then giving install instructions on how-to install the dependencies.

Documentation - Automated tests
---
No automated tests are provided, and no instructions are provided in the readme on howto
run manual tests of the library.

Documentation - Community guidelines
---
point 2) Report issues or problems with the software, and point 3) Seek support are to be addressed

Software paper - Authors
---
Author is listed, but the affiliation is mentioned as "None", if no affiliation exists, it would be more professional to mark the author as "independant developer"

Software paper - A statement of need
---
Same comment as before in the documenttation section, please provide a statement of need

Software paper - References
---
No DOIs are provided for the references

Additional
---
The text in the paper is of low quality in some areas, and the author might want to improve the text before publication, for example:

"Here are some ideas for future work:"
"These are the main resources besides the source code:"

Furthermore no comparison is made between the current clustering algorithm and known clustering
algorithm, how is this algorithm different from K-Means, Mean-Shift, DBSCAN, EM Clustering with
Gaussian Mixture Models (GMM) or Agglomerative Hierarchical Clustering.

To me it is not clear what advantages this clustering algorithm has, and it is not clear what
the author means by improving the "cluster's quality". How is this measured?

Furthermore, the phrase "they define a set of requirements cluster definitions must fulfill" is
very vague, and explanation is given how the user must do this.
