### 2018-09-18 #960 @volfpeter
Thanks for the comments! I'll start fixing these issues and get back to you when it's done.
