### 2018-09-22 83d0d79
added DOIs + fixed the van Dongen technical report reference
