+++ paper.bib
@@ -4,15 +4,17 @@
     Journal = "Journal of Statistical Mechanics: Theory and Experiment",
     Year = 2008,
     Volume = "2008",
-    Pages = "P10008"
+    Pages = "P10008",
+    DOI = "10.1088/1742-5468/2008/10/P10008"
 }
 
-@article{Dongen:2000,
+@techreport{Dongen:2000,
     Author = "van Dongen, S.",
     Title = "A cluster algorithm for graphs",
-    Journal = "Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam",
+    Institution = "National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam",
     Year = 2000,
-    Volume = "Technical Report INS-R0010"
+    Key = "INS-R0010",
+    URL = "https://micans.org/mcl/index.html"
 }
 
 @article{Girvan:2002,
@@ -21,7 +23,8 @@
     Journal = "PNAS",
     Year = 2002,
     Volume = "99 (12)",
-    Pages = "7821–7826"
+    Pages = "7821–7826",
+    DOI = "10.1073/pnas.122653799"
 }
 
 @article{Newman:2006,
@@ -30,7 +33,8 @@
     Journal = "Physical Review E",
     Year = 2006,
     Volume = "74",
-    Pages = "036104"
+    Pages = "036104",
+    DOI = "10.1103/PhysRevE.74.036104"
 }
 
 @article{Newman:2004,
@@ -39,7 +43,8 @@
     Journal = "Physical Review E",
     Year = 2004,
     Volume = "69",
-    Pages = "026113"
+    Pages = "026113",
+    DOI = "10.1103/PhysRevE.69.026113"
 }
 
 @article{Pons:2006,
@@ -48,7 +53,8 @@
     Journal = "Journal of Graph Algorithms and Applications",
     Year = 2006,
     Volume = "10",
-    Pages = "191–218"
+    Pages = "191–218",
+    DOI = "10.1007/11569596_31"
 }
 
 @article{Raghavan:2007,
@@ -57,7 +63,8 @@
     Journal = "Physical Review E",
     Year = 2007,
     Volume = "76",
-    Pages = "036106"
+    Pages = "036106",
+    DOI = "10.1103/PhysRevE.76.036106"
 }
 
 @article{Schaeffer:2007,
@@ -66,7 +73,8 @@
     Journal = "Computer Science Review",
     Year = 2007,
     Volume = "1",
-    Pages = "27-64"
+    Pages = "27-64",
+    DOI = "10.1016/j.cosrev.2007.05.001"
 }
 
 @article{Zachary:1977,
@@ -75,5 +83,6 @@
     Journal = "Journal of Anthropological Research",
     Year = 1977,
     Volume = "33",
-    Pages = "452--473"
+    Pages = "452--473",
+    DOI = "10.1086/jar.33.4.3629752"
 }
