### 2018-09-27 732fb69
updated setup.py + added pypi_readme.md
