### 2018-09-27 3c1203e
replaced README.rst with an improved version in markdown
