### 2018-09-27 2e7003c
use Independent developer as the affiliation + try a different kind of reference to cluster_example.png, hopefully it won't break the pdf build
