### 2018-09-27 #960 @volfpeter
@DannyArends I've updated the readme. Hopefully it's ok now, although i'm not sure you will be satisfied with the "statement of need" paragraph.
I'll update the paper soon, it's not ready yet.
