### 2018-10-01 #960 @DannyArends
@volfpeter Looks a lot better already, I check all the boxes, except the "automated tests", since @pjotrp mentioned that "Also tests are optional." I could just tick it if needed:

A couple of minor things:
In the README.md, please specify the language for the example text boxes, so that it will do code-highlighting, by specifying the language after the ``` to ```Python, or ```Bash:
```
from localclustering.definitions.connectivity import ConnectivityClusterDefinition
from localclustering.localengine import LocalClusterEngine
```
compared with (not the best example, but I think you'll get the idea):
```Python
from localclustering.definitions.connectivity import ConnectivityClusterDefinition
from localclustering.localengine import LocalClusterEngine
```

There is a missing space in the README.md: "a detailed descriptionand", as well as a minor spelling error: change 'in' to 'by' in the following sentence: "contact the repository owner in email"

Anyway, it looks good to me, just update the paper (if you want I could proofread it for you) and if @pjotrp confirms his "tests are optional" statement I will check the last box.

Danny
