### 2018-10-01 #960 @volfpeter
I've fixed the readme.

The reference that doesn't have a DOI is a technical report. I have found DOIs for related papers, but i wanted to cite the technical report because that's what i read back in the day.

Regarding the test, the only related thing in the repository is the `demo` module. It's easy to run if you went through the Getting started section of the readme. You will receive a very detailed error message saying you need to install the cairo library and its python binding (pycairo), then it will run.

One more thing: i've released a new version (0.11) that fixed the setup.py file, so that's the latest version right now.

Hopefully i will have time to finalize and commit the paper today.
