### 2018-10-01 #960 @DannyArends
@volfpeter I think it's more for regression/CI tests e.g. when using travis.ci

These kinds of tests allow you to know when a change to the code breaks the results compared to a previous version. I use it for the CTLmapping repository, every time I push to the master branch, the code is checked to make sure it can still install, and it runs tests to see if the mapping algorithm still produces the same output as before. See for example:

https://github.com/DannyArends/CTLmapping/blob/master/Rctl/tests/test_openmp.R

in this code my own openMP correlation algorithm is compared to the build-in correlation algorithm from R. If my algorithm produces different results then the R version

if(!all(res.ref == res.cor)) stop("ref != cor\n") # When results differ, stop

The test is stopped (stop is R lingo for error), and travis.ci will email me that the latest changes have broken the algorithm.

You could do something like this for the clustering, in which you calculate a base case, which you then test.

@pjotrp Is this a requirement or a nice to have feature for JOSS ?
