### 2018-10-01 #960 @DannyArends
Ok,
thanks for the confirmation, I checked all the boxes.
After the paper is updated, everything should be good 2 go

Danny
