### 2018-10-01 #960 @pjotrp
Yup. If there is no DOI there is none ;)
