### 2018-10-01 1a8d5c6
paper update
