### 2018-10-02 68b6dd4
Adding algorithms
