### 2018-10-02 d074548
minor changes to the first paragraph of the Algorithm section
