### 2018-10-02 0a169bd
Reworking the sentence
