+++ paper.md
@@ -63,7 +63,7 @@ Furthermore, for a cluster definition to be hierarchical, it must be able to adj
 
 ### Connectivity based cluster definition
 
-The connectivity based cluster definition is the default cluster definition implementation in this project that also happens to be a hierarchical one.
+The connectivity based cluster definition is the default (hierarchical) cluster definition implementation in this project.
 
 The cluster definition broadly works in the following way:
 
