### 2018-10-02 #960 @volfpeter
@DannyArends Thanks a lot for the fixes! I've merged the pull request.
@DannyArends @pjotrp Thanks for all the help during the pre-review and review process! It was a much better experience than i had expected.
