### 2018-10-03 #960 @arfon
> Excellent. The submission is slightly over 1000 words at 1200 words, but in this case I think it serves its purpose (@arfon is that OK?).

👍 LGTM
