### 2018-10-03 4ed802d
DOI badge and version update in setup.py
