### 2018-10-03 #960 @volfpeter
I've submitted the repository to Zenodo:
* DOI for "paper" release: 10.5281/zenodo.1443551
* DOI for all releases: 10.5281/zenodo.1443550

The repository version of the "paper" release is 0.12.1.
