### 2018-10-03 9a1d140
use the "all versions" DOI instead of the 0.12.1 DOI
