+++ README.md
@@ -1,4 +1,4 @@
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1443551.svg)](https://doi.org/10.5281/zenodo.1443551)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1443550.svg)](https://doi.org/10.5281/zenodo.1443550)
 
 # LocalClustering
 
